chore(auth): rename Admin files into Dashboard
